refactor (action): organizing pipes into folders and updated the design docs to reflect the changes.
