+++ projects/dsp-ui/src/lib/action/action.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormattedBooleanPipe } from './pipes/formatted-boolean.pipe';
-import { KeyPipe } from './pipes/key.pipe';
-import { KnoraDatePipe } from './pipes/knoradate.pipe';
-import { ReversePipe } from './pipes/reverse.pipe';
-import { SortByPipe } from './pipes/sort-by.pipe';
-import { TruncatePipe } from './pipes/truncate.pipe';
+import { FormattedBooleanPipe } from './pipes/formatting/formatted-boolean.pipe';
+import { KeyPipe } from './pipes/array-transformation/key.pipe';
+import { KnoraDatePipe } from './pipes/formatting/knoradate.pipe';
+import { ReversePipe } from './pipes/array-transformation/reverse.pipe';
+import { SortByPipe } from './pipes/array-transformation/sort-by.pipe';
+import { TruncatePipe } from './pipes/string-transformation/truncate.pipe';
 
 @NgModule({
   declarations: [
