+++ projects/dsp-ui/src/lib/action/pipes/array-transformation/reverse.pipe.ts
