+++ projects/dsp-ui/src/lib/action/pipes/string-transformation/truncate.pipe.spec.ts
