+++ projects/dsp-ui/src/lib/viewer/values/boolean-value/boolean-value.component.spec.ts
@@ -9,7 +9,7 @@ import { Component, OnInit, ViewChild, DebugElement } from '@angular/core';
 import { ReadBooleanValue, MockResource, UpdateBooleanValue, CreateBooleanValue } from '@knora/api';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FormattedBooleanPipe } from '../../../action/pipes/formatted-boolean.pipe';
+import { FormattedBooleanPipe } from '../../../action/pipes/formatting/formatted-boolean.pipe';
 
 /**
  * Test host component to simulate parent component.
